update: /multiplayer full re-factorization; components are not redirected to /multiplayer/controlCenter
